+++ lib/evm/precompiles/04-identity.js
@@ -1,6 +1,6 @@
 const utils = require('ethereumjs-util')
 const BN = utils.BN
-const error = require('../exceptions.js').ERROR
+const error = require('../../exceptions.js').ERROR
 const assert = require('assert')
 
 module.exports = function (opts) {
