+++ lib/evm/precompiles/06-ecadd.js
@@ -1,6 +1,6 @@
 const utils = require('ethereumjs-util')
-const ERROR = require('../exceptions.js').ERROR
 const BN = utils.BN
+const error = require('../../exceptions.js').ERROR
 const assert = require('assert')
 
 const bn128 = require('rustbn.js')
@@ -11,24 +11,23 @@ module.exports = function (opts) {
   let results = {}
   let inputData = opts.data
 
-  results.gasUsed = new BN(opts._common.param('gasPrices', 'ecMul'))
-
+  results.gasUsed = new BN(opts._common.param('gasPrices', 'ecAdd'))
   if (opts.gasLimit.lt(results.gasUsed)) {
     results.return = Buffer.alloc(0)
     results.exception = 0
     results.gasUsed = new BN(opts.gasLimit)
-    results.exceptionError = ERROR.OUT_OF_GAS
+    results.exceptionError = error.OUT_OF_GAS
     return results
   }
 
-  let returnData = bn128.mul(inputData)
+  let returnData = bn128.add(inputData)
 
-  // check ecmul success or failure by comparing the output length
+  // check ecadd success or failure by comparing the output length
   if (returnData.length !== 64) {
     results.return = Buffer.alloc(0)
     results.exception = 0
     results.gasUsed = new BN(opts.gasLimit)
-    results.exceptionError = ERROR.OUT_OF_GAS
+    results.exceptionError = error.OUT_OF_GAS
   } else {
     results.return = returnData
     results.exception = 1
