+++ lib/index.js
@@ -9,14 +9,14 @@ const fakeBlockchain = require('./fakeBlockChain.js')
 const BN = ethUtil.BN
 
 // require the precompiled contracts
-const num01 = require('./precompiled/01-ecrecover.js')
-const num02 = require('./precompiled/02-sha256.js')
-const num03 = require('./precompiled/03-ripemd160.js')
-const num04 = require('./precompiled/04-identity.js')
-const num05 = require('./precompiled/05-modexp.js')
-const num06 = require('./precompiled/06-ecadd.js')
-const num07 = require('./precompiled/07-ecmul.js')
-const num08 = require('./precompiled/08-ecpairing.js')
+const num01 = require('./evm/precompiles/01-ecrecover.js')
+const num02 = require('./evm/precompiles/02-sha256.js')
+const num03 = require('./evm/precompiles/03-ripemd160.js')
+const num04 = require('./evm/precompiles/04-identity.js')
+const num05 = require('./evm/precompiles/05-modexp.js')
+const num06 = require('./evm/precompiles/06-ecadd.js')
+const num07 = require('./evm/precompiles/07-ecmul.js')
+const num08 = require('./evm/precompiles/08-ecpairing.js')
 
 /**
  * VM Class, `new VM(opts)` creates a new VM object
